requirements.txt
beautifulsoup4==4.13.3
Flask==3.1.0
lxml==5.3.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.6.1
SQLAlchemy==2.0.37
psycopg2==2.9.10